Install Guardian from nuget, wiring into job.yml and respective proj files
