+++ eng/common/internal/Tools.csproj
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <!-- Copyright (c)  Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information. -->
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
@@ -10,11 +10,13 @@
     <PackageReference Remove="@(PackageReference)"/>
     <PackageReference Include="Microsoft.DotNet.IBCMerge" Version="$(MicrosoftDotNetIBCMergeVersion)" Condition="'$(UsingToolIbcOptimization)' == 'true'" />
     <PackageReference Include="Drop.App" Version="$(DropAppVersion)" ExcludeAssets="all" Condition="'$(UsingToolVisualStudioIbcTraining)' == 'true'"/>
+    <PackageReference Include ="Microsoft.Guardian.Cli" Version="0.0.38" Condition="'$(UsingToolSDL)' == 'true'"/>
   </ItemGroup>
   <PropertyGroup>
     <RestoreSources>
       https://devdiv.pkgs.visualstudio.com/_packaging/8f470c7e-ac49-4afe-a6ee-cf784e438b93/nuget/v3/index.json;
       https://devdiv.pkgs.visualstudio.com/_packaging/VS/nuget/v3/index.json;
+      https://securitytools.pkgs.visualstudio.com/_packaging/Guardian/nuget/v3/index.json
     </RestoreSources>
   </PropertyGroup>
