+++ eng/common/templates/job/job.yml
@@ -43,6 +43,9 @@ parameters:
   # Optional: enable sending telemetry
   enableTelemetry: false
 
+  # Optional: Include SDLAutomator task
+  enableSDL: false
+
   # Optional: define the helix repo for telemeetry (example: 'dotnet/arcade')
   helixRepo: ''
 
@@ -195,3 +198,29 @@ jobs:
         ArtifactName: AssetManifests
       continueOnError: ${{ parameters.continueOnError }}
       condition: and(succeeded(), eq(variables['_DotNetPublishToBlobFeed'], 'true'))
+
+  - ${{ if eq(parameters.enableSDL, 'true') }}:
+    - ${{ if and(eq(parameters.runAsPublic, 'false'), ne(variables['System.TeamProject'], 'public'), notin(variables['Build.Reason'], 'PullRequest')) }}:
+      - task: NuGetToolInstaller@0
+        displayName: Restore nuget.exe and add to PATH
+        inputs:
+            versionSpec: 4.4.1
+      - task: NuGetCommand@2
+         displayName: Restore internal tools
+         inputs:
+             command: restore
+             feedsToUse: config
+             restoreSolution: 'eng\common\internal\Tools.csproj'
+             nugetConfigPath: 'NuGet.config'
+             restoreDirectory: '$(Build.SourcesDirectory)\.packages'
+             env:
+                VSS_NUGET_EXTERNAL_FEED_ENDPOINTS: '{"endpointCredentials": [{"endpoint":"https://securitytools.pkgs.visualstudio.com/_packaging/Guardian/nuget/v3/index.json", "username":"dn-bot","password":"$(dn-bot-devdiv-nuget-feed-read)"}]}'
+                TargetGroup: $(_framework)
+      - task: PowerShell@2
+         displayName: Restore SDL Package
+         inputs:
+              filePath: eng\common\sdk-task.ps1
+              arguments: -task SDLAutomation -restore -msbuildEngine dotnet
+                  /p:PackageDirectory='$(Build.SourcesDirectory)/.packages'
+              condition: ${{ parameters.condition }}
+              continueOnError: ${{ parameters.continueOnError }}
