+++ src/Microsoft.DotNet.Arcade.Sdk/tools/SdkTasks/SDLAutomation.proj
@@ -1,0 +1,30 @@
+<!-- Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE file in the project root for more information. -->
+<Project Sdk="Microsoft.NET.Sdk" DefaultTargets="Execute">
+  <!--
+    Required variables:
+      PackagesDirectory                 Path to packages in the source directory.
+  -->
+  <PropertyGroup>
+    <TargetFramework>net472</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.DotNet.SDLAutomator" Version="$(MicrosoftDotNetSDLAutomatorVersion)" />
+  </ItemGroup>
+
+  <PropertyGroup>
+    <_MicrosoftDotNetSDLAutomatorTaskBaseDir>$(NuGetPackageRoot)microsoft.dotnet.sdlautomator\$(MicrosoftDotNetSDLAutomatorVersion)\tools\</_MicrosoftDotNetSDLAutomatorTaskBaseDir>
+    <_MicrosoftDotNetSDLAutomatorTaskDir>$(_MicrosoftDotNetSDLAutomatorTaskBaseDir)net472/</_MicrosoftDotNetSDLAutomatorTaskDir>
+    <_MicrosoftDotNetSDLAutomatorTaskDir Condition="'$(MSBuildRuntimeType)' == 'Core'">$(_MicrosoftDotNetSDLAutomatorTaskBaseDir)netcoreapp2.1/</_MicrosoftDotNetSDLAutomatorTaskDir>
+  </PropertyGroup>
+
+  <UsingTask TaskName="InstallGuardian" AssemblyFile="$(_MicrosoftDotNetSDLAutomatorTaskDir)Microsoft.DotNet.SDLAutomator.dll"/>
+
+  <Target Name="Execute">
+    <Error Text="The PackageDirectory property wasn't supplied."
+          Condition="'$(PackageDirectory)' == ''" />
+
+    <InstallGuardian PackageDirectory="$(PackageDirectory)"
+                     GuardianPackageVersion="0.0.38"/>
+  </Target>
+</Project>
