+++ src/Microsoft.DotNet.SDLAutomator/sdk/Sdk.props
@@ -1,0 +1,5 @@
+<!-- Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE file in the project root for more information. -->
+<Project>
+  <!-- Sdk.props file is required to be present for use as an Sdk, we don't have any props that we use though -->
+  <Import Project="$(MSBuildThisFileDirectory)..\build\Microsoft.DotNet.SDLAutomator.props" Condition="Exists('$(MSBuildThisFileDirectory)..\build\Microsoft.DotNet.SDLAutomator.props')" />
+</Project>
