+++ src/Microsoft.DotNet.SDLAutomator/sdk/Sdk.targets
@@ -1,0 +1,4 @@
+<!-- Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE file in the project root for more information. -->
+<Project>
+  <Import Project="$(MSBuildThisFileDirectory)..\build\Microsoft.DotNet.SDLAutomator.targets" />
+</Project>
